requirements.txt
django
djangorestframework
pyjwt